+++ src/components/layout/Header.tsx
@@ -3,23 +3,11 @@
 import { usePathname, useRouter } from "next/navigation";
 import PATH from "@/constants/PATH";
 import Image from "next/image";
-import { useState,useEffect } from "react";
-import NotificationDropdown from "@/components/notifications/NotificationDropdown";
-import getUserData from "@/supabase/getUserData";
-import useNotificationSubscription from "@/hooks/useNotificationSubscription";
-import {
-  DropdownMenu,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
+import NotificationButton from "@/components/layout/NotificationButton"; // 새로 만든 컴포넌트 임포트
 
-
-const Header = () => {
+const Header = () => {  
   const router = useRouter();
   const pathname = usePathname();
-  const [isDropdownOpen, setIsDropdownOpen] = useState(false);
-  const [newalarm, setNewalarm] = useState(false); // 새로운 알림 상태
-  const [userId, setUserId] = useState<string | null>(null);
-
 
   // 현재 랜딩 페이지인지 확인하기
   const isLandingPage = pathname === PATH.HOME;
@@ -43,26 +31,10 @@ const Header = () => {
     router.push(PATH.HOME);
   };
 
-  useEffect(() => {
-    const fetchId = async () => {
-      const userData = await getUserData();
-      setUserId(userData?.user?.id || null);
-    };
-
-    fetchId();
-  }, []);
-
-  useNotificationSubscription(userId, (payload) => {
-    console.log("새로운 알림 도착 (헤더) - Hook!", payload.new);
-    setNewalarm(true);
-  });
-
   return (
     <header className="fixed top-0 z-50 flex w-full items-center justify-between border-b bg-white px-4 py-4">
       {/* 뒤로가기 버튼 (랜딩 페이지에서는 숨김) */}
-      {isLandingPage ? (
-        <div className="h-5 w-5" /> // 공간만 유지
-      ) : (
+      {isLandingPage ? <div className="h-5 w-5" /> : (
         <button onClick={handleBack}>
           <Image
             src="/icons/chevron-left-solid 1.svg"
@@ -83,28 +55,8 @@ const Header = () => {
         ></Image>
       </button>
 
-      {/* 오른쪽 알림 아이콘 및 드롭다운 */}
-      <div className="relative">
-        <DropdownMenu open={isDropdownOpen} onOpenChange={setIsDropdownOpen}>
-          <DropdownMenuTrigger asChild>
-            <button className="relative ml-4">
-              {/* 조건부 이미지 렌더링 */}
-              <Image
-                src={
-                  newalarm ? "/icons/alarm1.svg" : "/icons/alarm.svg"
-                }
-                alt="알림로고"
-                height={24}
-                width={24}
-              />
-            </button>
-          </DropdownMenuTrigger>
-          <NotificationDropdown
-            open={isDropdownOpen}
-            onOpenChange={setIsDropdownOpen}
-          />
-        </DropdownMenu>
-      </div>
+      {/* 오른쪽 알림 버튼 */}
+      <NotificationButton />
     </header>
   );
 };
